imrove color functions

+++ Classes/FunObjc/Categories/UIColor+Fun.h
@@ -16,10 +16,10 @@
 #define GREEN [UIColor greenColor]
 #define BLUE [UIColor blueColor]
 #define RANDOM_COLOR [UIColor randomColor]
-UIColor* rgba(NSUInteger r, NSUInteger g, NSUInteger b, CGFloat a);
-UIColor* rgb(NSUInteger r, NSUInteger g, NSUInteger b);
-UIColor* hsva(NSUInteger h, NSUInteger s, NSUInteger v, CGFloat a);
-UIColor* hsv(NSUInteger h, NSUInteger s, NSUInteger v);
+UIColor* rgba(CGFloat r, CGFloat g, CGFloat b, CGFloat a);
+UIColor* rgb(CGFloat r, CGFloat g, CGFloat b);
+UIColor* hsva(CGFloat h, CGFloat s, CGFloat v, CGFloat a);
+UIColor* hsv(CGFloat h, CGFloat s, CGFloat v);
 
 #define LIGHT_GRAY rgb(230,230,230)
 #define STEELBLUE rgb(70,130,180)
