+++ Classes/FunObjc/Categories/UIColor+Fun.m
@@ -8,16 +8,16 @@
 
 #import "UIColor+Fun.h"
 
-UIColor* rgba(NSUInteger r, NSUInteger g, NSUInteger b, CGFloat a) {
+UIColor* rgba(CGFloat r, CGFloat g, CGFloat b, CGFloat a) {
     return [UIColor colorWithRed:r/255.0 green:g/255.0 blue:b/255.0 alpha:a];
 }
-UIColor* rgb(NSUInteger r, NSUInteger g, NSUInteger b) {
+UIColor* rgb(CGFloat r, CGFloat g, CGFloat b) {
     return rgba(r, g, b, 1.0);
 }
-UIColor* hsva(NSUInteger h, NSUInteger s, NSUInteger v, CGFloat a) {
+UIColor* hsva(CGFloat h, CGFloat s, CGFloat v, CGFloat a) {
     return [UIColor colorWithHue:h/360.0 saturation:s/100.0 brightness:v/100.0 alpha:a/100.0];
 }
-UIColor* hsv(NSUInteger h, NSUInteger s, NSUInteger v) {
+UIColor* hsv(CGFloat h, CGFloat s, CGFloat v) {
     return hsva(h, s, v, 100.0);
 }
